[Improvement][api] Introduce interface for monitor, processinstance, queue service
