+++ dolphinscheduler-api/src/test/java/org/apache/dolphinscheduler/api/service/MonitorServiceTest.java
@@ -17,12 +17,18 @@
 package org.apache.dolphinscheduler.api.service;
 
 import org.apache.dolphinscheduler.api.enums.Status;
+import org.apache.dolphinscheduler.api.service.impl.MonitorServiceImpl;
 import org.apache.dolphinscheduler.common.Constants;
 import org.apache.dolphinscheduler.common.enums.DbType;
 import org.apache.dolphinscheduler.common.model.Server;
 import org.apache.dolphinscheduler.common.utils.CollectionUtils;
 import org.apache.dolphinscheduler.dao.MonitorDBDao;
 import org.apache.dolphinscheduler.dao.entity.MonitorRecord;
+
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Map;
+
 import org.junit.Assert;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -33,40 +39,39 @@ import org.mockito.junit.MockitoJUnitRunner;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Map;
-
 @RunWith(MockitoJUnitRunner.class)
 public class MonitorServiceTest {
 
     private static final Logger logger = LoggerFactory.getLogger(MonitorServiceTest.class);
 
     @InjectMocks
-    private MonitorService monitorService;
+    private MonitorServiceImpl monitorService;
+
     @Mock
     private MonitorDBDao monitorDBDao;
 
 
     @Test
-    public  void testQueryDatabaseState(){
+    public void testQueryDatabaseState() {
 
         Mockito.when(monitorDBDao.queryDatabaseState()).thenReturn(getList());
-        Map<String,Object> result = monitorService.queryDatabaseState(null);
+        Map<String, Object> result = monitorService.queryDatabaseState(null);
         logger.info(result.toString());
-        Assert.assertEquals(Status.SUCCESS,result.get(Constants.STATUS));
+        Assert.assertEquals(Status.SUCCESS, result.get(Constants.STATUS));
         List<MonitorRecord> monitorRecordList = (List<MonitorRecord>) result.get(Constants.DATA_LIST);
         Assert.assertTrue(CollectionUtils.isNotEmpty(monitorRecordList));
     }
+
     @Test
-    public  void testQueryMaster(){
+    public void testQueryMaster() {
         //TODO need zk
 //        Map<String,Object> result = monitorService.queryMaster(null);
 //        logger.info(result.toString());
 //        Assert.assertEquals(Status.SUCCESS,result.get(Constants.STATUS));
     }
+
     @Test
-    public  void testQueryZookeeperState(){
+    public void testQueryZookeeperState() {
         //TODO need zk
 //        Map<String,Object> result = monitorService.queryZookeeperState(null);
 //        logger.info(result.toString());
@@ -74,25 +79,25 @@ public class MonitorServiceTest {
     }
 
     @Test
-    public  void testGetServerListFromZK(){
+    public void testGetServerListFromZK() {
         //TODO need zk
 //        List<Server> serverList = monitorService.getServerListFromZK(true);
 //        logger.info(serverList.toString());
     }
 
-    private List<MonitorRecord> getList(){
+    private List<MonitorRecord> getList() {
         List<MonitorRecord> monitorRecordList = new ArrayList<>();
         monitorRecordList.add(getEntity());
         return monitorRecordList;
     }
 
-    private MonitorRecord getEntity(){
-        MonitorRecord monitorRecord = new  MonitorRecord();
+    private MonitorRecord getEntity() {
+        MonitorRecord monitorRecord = new MonitorRecord();
         monitorRecord.setDbType(DbType.MYSQL);
         return monitorRecord;
     }
 
-    private List<Server> getServerList(){
+    private List<Server> getServerList() {
         List<Server> servers = new ArrayList<>();
         servers.add(new Server());
         return servers;
