+++ dolphinscheduler-api/src/test/java/org/apache/dolphinscheduler/api/service/QueueServiceTest.java
@@ -16,9 +16,8 @@
  */
 package org.apache.dolphinscheduler.api.service;
 
-import com.baomidou.mybatisplus.core.metadata.IPage;
-import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
 import org.apache.dolphinscheduler.api.enums.Status;
+import org.apache.dolphinscheduler.api.service.impl.QueueServiceImpl;
 import org.apache.dolphinscheduler.api.utils.PageInfo;
 import org.apache.dolphinscheduler.api.utils.Result;
 import org.apache.dolphinscheduler.common.Constants;
@@ -28,6 +27,11 @@ import org.apache.dolphinscheduler.dao.entity.Queue;
 import org.apache.dolphinscheduler.dao.entity.User;
 import org.apache.dolphinscheduler.dao.mapper.QueueMapper;
 import org.apache.dolphinscheduler.dao.mapper.UserMapper;
+
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Map;
+
 import org.junit.After;
 import org.junit.Assert;
 import org.junit.Before;
@@ -40,9 +44,8 @@ import org.mockito.junit.MockitoJUnitRunner;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Map;
+import com.baomidou.mybatisplus.core.metadata.IPage;
+import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
 
 @RunWith(MockitoJUnitRunner.class)
 public class QueueServiceTest {
@@ -50,11 +53,14 @@ public class QueueServiceTest {
     private static final Logger logger = LoggerFactory.getLogger(QueueServiceTest.class);
 
     @InjectMocks
-    private QueueService queueService;
+    private QueueServiceImpl queueService;
+
     @Mock
     private QueueMapper queueMapper;
+
     @Mock
     private UserMapper userMapper;
+
     private String queueName = "QueueServiceTest";
 
     @Before
@@ -63,50 +69,53 @@ public class QueueServiceTest {
 
 
     @After
-    public void after(){
+    public void after() {
     }
 
     @Test
-    public void testQueryList(){
+    public void testQueryList() {
 
         Mockito.when(queueMapper.selectList(null)).thenReturn(getQueueList());
         Map<String, Object> result = queueService.queryList(getLoginUser());
         logger.info(result.toString());
-        List<Queue> queueList  = (List<Queue>) result.get(Constants.DATA_LIST);
+        List<Queue> queueList = (List<Queue>) result.get(Constants.DATA_LIST);
         Assert.assertTrue(CollectionUtils.isNotEmpty(queueList));
 
     }
-    @Test
-    public void testQueryListPage(){
 
-        IPage<Queue> page = new Page<>(1,10);
+    @Test
+    public void testQueryListPage() {
+
+        IPage<Queue> page = new Page<>(1, 10);
         page.setTotal(1L);
         page.setRecords(getQueueList());
         Mockito.when(queueMapper.queryQueuePaging(Mockito.any(Page.class), Mockito.eq(queueName))).thenReturn(page);
-        Map<String, Object> result = queueService.queryList(getLoginUser(),queueName,1,10);
+        Map<String, Object> result = queueService.queryList(getLoginUser(), queueName, 1, 10);
         logger.info(result.toString());
-        PageInfo<Queue>  pageInfo = (PageInfo<Queue>) result.get(Constants.DATA_LIST);
+        PageInfo<Queue> pageInfo = (PageInfo<Queue>) result.get(Constants.DATA_LIST);
         Assert.assertTrue(CollectionUtils.isNotEmpty(pageInfo.getLists()));
     }
+
     @Test
-    public void testCreateQueue(){
+    public void testCreateQueue() {
 
         // queue is null
-        Map<String, Object> result = queueService.createQueue(getLoginUser(),null,queueName);
+        Map<String, Object> result = queueService.createQueue(getLoginUser(), null, queueName);
         logger.info(result.toString());
-        Assert.assertEquals(Status.REQUEST_PARAMS_NOT_VALID_ERROR,result.get(Constants.STATUS));
+        Assert.assertEquals(Status.REQUEST_PARAMS_NOT_VALID_ERROR, result.get(Constants.STATUS));
         // queueName is null
-        result = queueService.createQueue(getLoginUser(),queueName,null);
+        result = queueService.createQueue(getLoginUser(), queueName, null);
         logger.info(result.toString());
-        Assert.assertEquals(Status.REQUEST_PARAMS_NOT_VALID_ERROR,result.get(Constants.STATUS));
+        Assert.assertEquals(Status.REQUEST_PARAMS_NOT_VALID_ERROR, result.get(Constants.STATUS));
         // correct
-        result = queueService.createQueue(getLoginUser(),queueName,queueName);
+        result = queueService.createQueue(getLoginUser(), queueName, queueName);
         logger.info(result.toString());
-        Assert.assertEquals(Status.SUCCESS,result.get(Constants.STATUS));
+        Assert.assertEquals(Status.SUCCESS, result.get(Constants.STATUS));
 
     }
+
     @Test
-    public void testUpdateQueue(){
+    public void testUpdateQueue() {
 
         Mockito.when(queueMapper.selectById(1)).thenReturn(getQueue());
         Mockito.when(queueMapper.queryAllQueueList("test", null)).thenReturn(getQueueList());
@@ -114,65 +123,66 @@ public class QueueServiceTest {
         Mockito.when(userMapper.queryUserListByQueue(queueName)).thenReturn(getUserList());
 
         // not exist
-        Map<String, Object> result = queueService.updateQueue(getLoginUser(),0,"queue",queueName);
+        Map<String, Object> result = queueService.updateQueue(getLoginUser(), 0, "queue", queueName);
         logger.info(result.toString());
-        Assert.assertEquals(Status.QUEUE_NOT_EXIST.getCode(),((Status)result.get(Constants.STATUS)).getCode());
+        Assert.assertEquals(Status.QUEUE_NOT_EXIST.getCode(), ((Status) result.get(Constants.STATUS)).getCode());
         //no need update
-        result = queueService.updateQueue(getLoginUser(),1,queueName,queueName);
+        result = queueService.updateQueue(getLoginUser(), 1, queueName, queueName);
         logger.info(result.toString());
-        Assert.assertEquals(Status.NEED_NOT_UPDATE_QUEUE.getCode(),((Status)result.get(Constants.STATUS)).getCode());
+        Assert.assertEquals(Status.NEED_NOT_UPDATE_QUEUE.getCode(), ((Status) result.get(Constants.STATUS)).getCode());
         //queue exist
-        result = queueService.updateQueue(getLoginUser(),1,"test",queueName);
+        result = queueService.updateQueue(getLoginUser(), 1, "test", queueName);
         logger.info(result.toString());
-        Assert.assertEquals(Status.QUEUE_VALUE_EXIST.getCode(),((Status)result.get(Constants.STATUS)).getCode());
+        Assert.assertEquals(Status.QUEUE_VALUE_EXIST.getCode(), ((Status) result.get(Constants.STATUS)).getCode());
         // queueName exist
-        result = queueService.updateQueue(getLoginUser(),1,"test1","test");
+        result = queueService.updateQueue(getLoginUser(), 1, "test1", "test");
         logger.info(result.toString());
-        Assert.assertEquals(Status.QUEUE_NAME_EXIST.getCode(),((Status)result.get(Constants.STATUS)).getCode());
+        Assert.assertEquals(Status.QUEUE_NAME_EXIST.getCode(), ((Status) result.get(Constants.STATUS)).getCode());
         //success
-        result = queueService.updateQueue(getLoginUser(),1,"test1","test1");
+        result = queueService.updateQueue(getLoginUser(), 1, "test1", "test1");
         logger.info(result.toString());
-        Assert.assertEquals(Status.SUCCESS.getCode(),((Status)result.get(Constants.STATUS)).getCode());
+        Assert.assertEquals(Status.SUCCESS.getCode(), ((Status) result.get(Constants.STATUS)).getCode());
 
     }
+
     @Test
-    public void testVerifyQueue(){
+    public void testVerifyQueue() {
 
         Mockito.when(queueMapper.queryAllQueueList(queueName, null)).thenReturn(getQueueList());
         Mockito.when(queueMapper.queryAllQueueList(null, queueName)).thenReturn(getQueueList());
 
         //queue null
-        Result result = queueService.verifyQueue(null,queueName);
+        Result result = queueService.verifyQueue(null, queueName);
         logger.info(result.toString());
         Assert.assertEquals(result.getCode().intValue(), Status.REQUEST_PARAMS_NOT_VALID_ERROR.getCode());
 
         //queueName null
-        result = queueService.verifyQueue(queueName,null);
+        result = queueService.verifyQueue(queueName, null);
         logger.info(result.toString());
         Assert.assertEquals(result.getCode().intValue(), Status.REQUEST_PARAMS_NOT_VALID_ERROR.getCode());
 
         //exist queueName
-        result = queueService.verifyQueue(queueName,queueName);
+        result = queueService.verifyQueue(queueName, queueName);
         logger.info(result.toString());
         Assert.assertEquals(result.getCode().intValue(), Status.QUEUE_NAME_EXIST.getCode());
 
         //exist queue
-        result = queueService.verifyQueue(queueName,"test");
+        result = queueService.verifyQueue(queueName, "test");
         logger.info(result.toString());
         Assert.assertEquals(result.getCode().intValue(), Status.QUEUE_VALUE_EXIST.getCode());
 
         // success
-        result = queueService.verifyQueue("test","test");
+        result = queueService.verifyQueue("test", "test");
         logger.info(result.toString());
         Assert.assertEquals(result.getCode().intValue(), Status.SUCCESS.getCode());
 
 
     }
+
     /**
      * create admin user
-     * @return
      */
-    private User getLoginUser(){
+    private User getLoginUser() {
 
         User loginUser = new User();
         loginUser.setUserType(UserType.ADMIN_USER);
@@ -180,7 +190,7 @@ public class QueueServiceTest {
         return loginUser;
     }
 
-    private List<User> getUserList(){
+    private List<User> getUserList() {
         List<User> list = new ArrayList<>();
         list.add(getLoginUser());
         return list;
@@ -189,9 +199,8 @@ public class QueueServiceTest {
 
     /**
      * get queue
-     * @return
      */
-    private Queue getQueue(){
+    private Queue getQueue() {
         Queue queue = new Queue();
         queue.setId(1);
         queue.setQueue(queueName);
@@ -199,7 +208,7 @@ public class QueueServiceTest {
         return queue;
     }
 
-    private List<Queue> getQueueList(){
+    private List<Queue> getQueueList() {
         List<Queue> queueList = new ArrayList<>();
         queueList.add(getQueue());
         return queueList;
